plant_disease/requirements.txt
flask
onnxruntime
numpy
pillow